# Speed-up Genomic Benchmarks dataloader